# Optimizations for DiscreteTrajectory