--- a/physics/discrete_trajectory_segment_body.hpp
+++ b/physics/discrete_trajectory_segment_body.hpp
@@ -107,55 +107,22 @@ void DiscreteTrajectorySegment<Frame>::clear() {
 template<typename Frame>
 typename DiscreteTrajectorySegment<Frame>::iterator
 DiscreteTrajectorySegment<Frame>::find(Instant const& t) const {
-  auto const it = FindImpl(t);
+  auto const it = FindOrNullopt(t);
   return it.has_value() ? *it : end();
-}
-
-template <typename Frame>
-std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
-DiscreteTrajectorySegment<Frame>::FindImpl(Instant const& t) const {
-  auto const it = timeline_.find(t);
-  if (it == timeline_.end()) {
-    return std::nullopt;
-  } else {
-    return iterator(self_, it);
-  }
 }
 
 template<typename Frame>
 typename DiscreteTrajectorySegment<Frame>::iterator
 DiscreteTrajectorySegment<Frame>::lower_bound(Instant const& t) const {
-  auto const it = LowerBoundImpl(t);
+  auto const it = LowerBoundOrNullopt(t);
   return it.has_value() ? *it : end();
-}
-
-template<typename Frame>
-std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
-DiscreteTrajectorySegment<Frame>::LowerBoundImpl(Instant const& t) const {
-  auto const it = timeline_.lower_bound(t);
-  if (it == timeline_.end()) {
-    return std::nullopt;
-  } else {
-    return iterator(self_, it);
-  }
 }
 
 template<typename Frame>
 typename DiscreteTrajectorySegment<Frame>::iterator
 DiscreteTrajectorySegment<Frame>::upper_bound(Instant const& t) const {
-  auto const it = UpperBoundImpl(t);
+  auto const it = UpperBoundOrNullopt(t);
   return it.has_value() ? *it : end();
-}
-
-template<typename Frame>
-std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
-DiscreteTrajectorySegment<Frame>::UpperBoundImpl(Instant const& t) const {
-  auto const it = timeline_.upper_bound(t);
-  if (it == timeline_.end()) {
-    return std::nullopt;
-  } else {
-    return iterator(self_, it);
-  }
 }
 
 template<typename Frame>
@@ -345,6 +312,39 @@ DiscreteTrajectorySegment<Frame>::ReadFromMessage(
   }
 
   return segment;
+}
+
+template <typename Frame>
+std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
+DiscreteTrajectorySegment<Frame>::FindOrNullopt(Instant const& t) const {
+  auto const it = timeline_.find(t);
+  if (it == timeline_.end()) {
+    return std::nullopt;
+  } else {
+    return iterator(self_, it);
+  }
+}
+
+template<typename Frame>
+std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
+DiscreteTrajectorySegment<Frame>::LowerBoundOrNullopt(Instant const& t) const {
+  auto const it = timeline_.lower_bound(t);
+  if (it == timeline_.end()) {
+    return std::nullopt;
+  } else {
+    return iterator(self_, it);
+  }
+}
+
+template<typename Frame>
+std::optional<typename DiscreteTrajectorySegment<Frame>::iterator>
+DiscreteTrajectorySegment<Frame>::UpperBoundOrNullopt(Instant const& t) const {
+  auto const it = timeline_.upper_bound(t);
+  if (it == timeline_.end()) {
+    return std::nullopt;
+  } else {
+    return iterator(self_, it);
+  }
 }
 
 template<typename Frame>