--- a/physics/discrete_trajectory_test.cpp
+++ b/physics/discrete_trajectory_test.cpp
@@ -137,6 +137,28 @@ TEST_F(DiscreteTrajectoryTest, BackFront) {
   EXPECT_EQ(t0_ + 14 * Second, trajectory.back().time);
 }
 
+TEST_F(DiscreteTrajectoryTest, FrontEmpty) {
+  // Construct a non-empty trajectory with an empty front segment.
+  DegreesOfFreedom<World> const dof(
+      World::origin, Velocity<World>({1 * Metre / Second, 0 * Metre / Second,
+                                      0 * Metre / Second}));
+  DiscreteTrajectory<World> trajectory;
+  ASSERT_TRUE(trajectory.Append(t0_, dof).ok());
+  trajectory.AttachSegments(MakeTrajectory(t0_, dof));
+
+  Instant const t1 = t0_ + 1 * Second;
+  trajectory.ForgetBefore(t1);
+  ASSERT_FALSE(trajectory.empty());
+  ASSERT_TRUE(trajectory.segments().front().empty());
+
+  // Verify that begin() and front() behave as expected.
+  EXPECT_EQ(trajectory.front().time, t1);
+  EXPECT_EQ(trajectory.begin()->time, t1);
+
+  EXPECT_EQ(trajectory.segments().front().front().time, t1);
+  EXPECT_EQ(trajectory.segments().front().begin()->time, t1);
+}
+
 TEST_F(DiscreteTrajectoryTest, IterateForward) {
   auto const trajectory = MakeTrajectory();
   std::vector<Instant> times;