To `front` and `end`.
Also, renamed the existing Front/Back/etc benchmarks as BM_DiscreteTrajectorySegmentFoo (because they measure the performance of segment functions), and created new benchmarks that measure the corresponding functions on the entire trajectory.

Before optimizations, but including benchmark reconfiguration (#5b6df36):
```
Benchmark                                                          Time             CPU   Iterations
----------------------------------------------------------------------------------------------------
BM_DiscreteTrajectoryFront                                      2.06 ns         2.06 ns    335252253
BM_DiscreteTrajectoryBack                                       18.5 ns         18.5 ns     37876943
BM_DiscreteTrajectoryBegin                                      1.22 ns         1.22 ns    572222676
BM_DiscreteTrajectoryEnd                                        7.84 ns         7.83 ns     89294826
BM_DiscreteTrajectoryTMin                                       2.34 ns         2.34 ns    298857936
BM_DiscreteTrajectoryTMax                                       4.41 ns         4.41 ns    158582724
BM_DiscreteTrajectorySegmentFront                              0.316 ns        0.316 ns   1000000000
BM_DiscreteTrajectorySegmentBack                               0.590 ns        0.590 ns   1000000000
BM_DiscreteTrajectorySegmentBegin                              0.789 ns        0.789 ns    884765600
BM_DiscreteTrajectorySegmentEnd                                 6.32 ns         6.31 ns    110878794
BM_DiscreteTrajectorySegmentTMin                                1.90 ns         1.90 ns    369151584
BM_DiscreteTrajectorySegmentTMax                                4.42 ns         4.42 ns    158637350
BM_DiscreteTrajectoryCreateDestroy/8                            1323 ns         1322 ns       528753
BM_DiscreteTrajectoryCreateDestroy/64                           4301 ns         4299 ns       163910
BM_DiscreteTrajectoryCreateDestroy/512                         27396 ns        27381 ns        25575
BM_DiscreteTrajectoryCreateDestroy/1024                        54314 ns        54277 ns        12885
BM_DiscreteTrajectoryIterate/8                                  21.6 ns         21.5 ns     32600900
BM_DiscreteTrajectoryIterate/64                                  152 ns          151 ns      4615801
BM_DiscreteTrajectoryIterate/512                                1180 ns         1179 ns       592934
BM_DiscreteTrajectoryIterate/1024                               2326 ns         2325 ns       299891
BM_DiscreteTrajectoryReverseIterate/8                           19.5 ns         19.5 ns     35863412
BM_DiscreteTrajectoryReverseIterate/64                           181 ns          181 ns      3881320
BM_DiscreteTrajectoryReverseIterate/512                         1490 ns         1489 ns       469358
BM_DiscreteTrajectoryReverseIterate/1024                        2940 ns         2939 ns       238825
BM_DiscreteTrajectoryFind/8                                     95.5 ns         95.4 ns      7303535
BM_DiscreteTrajectoryFind/64                                     121 ns          121 ns      5787994
BM_DiscreteTrajectoryFind/512                                    139 ns          139 ns      4949515
BM_DiscreteTrajectoryFind/1024                                   145 ns          144 ns      4830318
BM_DiscreteTrajectoryLowerBound/8                               58.5 ns         58.4 ns     11962540
BM_DiscreteTrajectoryLowerBound/64                              73.3 ns         73.2 ns      9503122
BM_DiscreteTrajectoryLowerBound/512                             90.3 ns         90.3 ns      7786256
BM_DiscreteTrajectoryLowerBound/1024                            94.1 ns         94.0 ns      7421386
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomExact              8.51 ns         8.51 ns     82096992
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomInterpolated       20.6 ns         20.6 ns     34142844
```

After optimizations (#dc232bd):
```
Benchmark                                                          Time             CPU   Iterations
----------------------------------------------------------------------------------------------------
BM_DiscreteTrajectoryFront                                     0.448 ns        0.448 ns   1000000000
BM_DiscreteTrajectoryBack                                       3.67 ns         3.67 ns    190162100
BM_DiscreteTrajectoryBegin                                      1.30 ns         1.30 ns    575232147
BM_DiscreteTrajectoryEnd                                       0.631 ns        0.631 ns   1000000000
BM_DiscreteTrajectoryTMin                                       2.39 ns         2.39 ns    297638030
BM_DiscreteTrajectoryTMax                                       2.63 ns         2.63 ns    269664306
BM_DiscreteTrajectorySegmentFront                              0.316 ns        0.316 ns   1000000000
BM_DiscreteTrajectorySegmentBack                               0.589 ns        0.589 ns   1000000000
BM_DiscreteTrajectorySegmentBegin                              0.789 ns        0.788 ns    886019872
BM_DiscreteTrajectorySegmentEnd                                 6.32 ns         6.32 ns    110255320
BM_DiscreteTrajectorySegmentTMin                                1.89 ns         1.89 ns    368419114
BM_DiscreteTrajectorySegmentTMax                                2.22 ns         2.22 ns    316769315
BM_DiscreteTrajectoryCreateDestroy/8                            1280 ns         1279 ns       537403
BM_DiscreteTrajectoryCreateDestroy/64                           4220 ns         4217 ns       167352
BM_DiscreteTrajectoryCreateDestroy/512                         26729 ns        26713 ns        25983
BM_DiscreteTrajectoryCreateDestroy/1024                        53650 ns        53619 ns        13030
BM_DiscreteTrajectoryIterate/8                                  20.8 ns         20.8 ns     33638159
BM_DiscreteTrajectoryIterate/64                                  143 ns          143 ns      4878627
BM_DiscreteTrajectoryIterate/512                                1201 ns         1201 ns       579005
BM_DiscreteTrajectoryIterate/1024                               2398 ns         2397 ns       291498
BM_DiscreteTrajectoryReverseIterate/8                           19.2 ns         19.2 ns     36514055
BM_DiscreteTrajectoryReverseIterate/64                           181 ns          181 ns      3841763
BM_DiscreteTrajectoryReverseIterate/512                         1492 ns         1491 ns       468751
BM_DiscreteTrajectoryReverseIterate/1024                        2943 ns         2941 ns       238100
BM_DiscreteTrajectoryFind/8                                     79.3 ns         79.2 ns      8869629
BM_DiscreteTrajectoryFind/64                                    96.3 ns         96.3 ns      7271443
BM_DiscreteTrajectoryFind/512                                    117 ns          117 ns      5980912
BM_DiscreteTrajectoryFind/1024                                   118 ns          118 ns      5954660
BM_DiscreteTrajectoryLowerBound/8                               58.9 ns         58.8 ns     11972566
BM_DiscreteTrajectoryLowerBound/64                              73.2 ns         73.1 ns      9578151
BM_DiscreteTrajectoryLowerBound/512                             87.3 ns         87.3 ns      7903889
BM_DiscreteTrajectoryLowerBound/1024                            93.8 ns         93.8 ns      7445066
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomExact              8.56 ns         8.56 ns     82251337
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomInterpolated       18.4 ns         18.4 ns     38056715
```

Diff:
```
Benchmark                                                                   Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------------------------------------
BM_DiscreteTrajectoryFront                                               -0.7831         -0.7831             2             0             2             0
BM_DiscreteTrajectoryBack                                                -0.8015         -0.8015            19             4            19             4
BM_DiscreteTrajectoryBegin                                               +0.0697         +0.0695             1             1             1             1
BM_DiscreteTrajectoryEnd                                                 -0.9195         -0.9195             8             1             8             1
BM_DiscreteTrajectoryTMin                                                +0.0219         +0.0219             2             2             2             2
BM_DiscreteTrajectoryTMax                                                -0.4038         -0.4037             4             3             4             3
BM_DiscreteTrajectorySegmentFront                                        -0.0001         -0.0003             0             0             0             0
BM_DiscreteTrajectorySegmentBack                                         -0.0022         -0.0022             1             1             1             1
BM_DiscreteTrajectorySegmentBegin                                        -0.0009         -0.0011             1             1             1             1
BM_DiscreteTrajectorySegmentEnd                                          +0.0008         +0.0009             6             6             6             6
BM_DiscreteTrajectorySegmentTMin                                         -0.0022         -0.0024             2             2             2             2
BM_DiscreteTrajectorySegmentTMax                                         -0.4973         -0.4974             4             2             4             2
BM_DiscreteTrajectoryCreateDestroy/8                                     -0.0326         -0.0327          1323          1280          1322          1279
BM_DiscreteTrajectoryCreateDestroy/64                                    -0.0188         -0.0189          4301          4220          4299          4217
BM_DiscreteTrajectoryCreateDestroy/512                                   -0.0243         -0.0244         27396         26729         27381         26713
BM_DiscreteTrajectoryCreateDestroy/1024                                  -0.0122         -0.0121         54314         53650         54277         53619
BM_DiscreteTrajectoryIterate/8                                           -0.0338         -0.0335            22            21            22            21
BM_DiscreteTrajectoryIterate/64                                          -0.0559         -0.0558           152           143           151           143
BM_DiscreteTrajectoryIterate/512                                         +0.0183         +0.0182          1180          1201          1179          1201
BM_DiscreteTrajectoryIterate/1024                                        +0.0312         +0.0310          2326          2398          2325          2397
BM_DiscreteTrajectoryReverseIterate/8                                    -0.0144         -0.0145            20            19            20            19
BM_DiscreteTrajectoryReverseIterate/64                                   -0.0006         -0.0008           181           181           181           181
BM_DiscreteTrajectoryReverseIterate/512                                  +0.0009         +0.0009          1490          1492          1489          1491
BM_DiscreteTrajectoryReverseIterate/1024                                 +0.0009         +0.0007          2940          2943          2939          2941
BM_DiscreteTrajectoryFind/8                                              -0.1697         -0.1698            95            79            95            79
BM_DiscreteTrajectoryFind/64                                             -0.2051         -0.2050           121            96           121            96
BM_DiscreteTrajectoryFind/512                                            -0.1617         -0.1618           139           117           139           117
BM_DiscreteTrajectoryFind/1024                                           -0.1836         -0.1838           145           118           144           118
BM_DiscreteTrajectoryLowerBound/8                                        +0.0070         +0.0070            58            59            58            59
BM_DiscreteTrajectoryLowerBound/64                                       -0.0015         -0.0016            73            73            73            73
BM_DiscreteTrajectoryLowerBound/512                                      -0.0333         -0.0332            90            87            90            87
BM_DiscreteTrajectoryLowerBound/1024                                     -0.0028         -0.0029            94            94            94            94
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomExact                       +0.0060         +0.0059             9             9             9             9
BM_DiscreteTrajectoryEvaluateDegreesOfFreedomInterpolated                -0.1061         -0.1062            21            18            21            18
```

Note that the TMax benchmarks appear to be noisy (TMax is not affected by this change).