@rnlahaye 2022-01-23
I optimized `DiscreteTrajectory::find` and friends somewhat by bypassing the call to `DiscreteTrajectorySegment::end`.

Old:
```
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_DiscreteTrajectoryFind/8                97.9 ns         97.4 ns      7185752
BM_DiscreteTrajectoryFind/64                122 ns          122 ns      5721619
BM_DiscreteTrajectoryFind/512               140 ns          139 ns      4987638
BM_DiscreteTrajectoryFind/1024              143 ns          142 ns      4952456
BM_DiscreteTrajectoryLowerBound/8          59.6 ns         59.5 ns     11828720
BM_DiscreteTrajectoryLowerBound/64         74.2 ns         74.1 ns      9457287
BM_DiscreteTrajectoryLowerBound/512        92.4 ns         92.3 ns      7593262
BM_DiscreteTrajectoryLowerBound/1024       95.1 ns         95.0 ns      7329996
```

New:
```
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_DiscreteTrajectoryFind/8                22.7 ns         22.7 ns     30914222
BM_DiscreteTrajectoryFind/64               34.4 ns         34.4 ns     20379644
BM_DiscreteTrajectoryFind/512              51.5 ns         51.4 ns     13518472
BM_DiscreteTrajectoryFind/1024             52.5 ns         52.4 ns     13293832
BM_DiscreteTrajectoryLowerBound/8          36.9 ns         36.8 ns     18994847
BM_DiscreteTrajectoryLowerBound/64         53.7 ns         53.7 ns     13234265
BM_DiscreteTrajectoryLowerBound/512        74.4 ns         74.4 ns      9413410
BM_DiscreteTrajectoryLowerBound/1024       76.3 ns         76.2 ns      9148533
```

Delta:
```
Benchmark                                              Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------------------------
BM_DiscreteTrajectoryFind/8                         -0.7685         -0.7674            98            23            97            23
BM_DiscreteTrajectoryFind/64                        -0.7181         -0.7181           122            34           122            34
BM_DiscreteTrajectoryFind/512                       -0.6311         -0.6312           140            51           139            51
BM_DiscreteTrajectoryFind/1024                      -0.6318         -0.6318           143            52           142            52
BM_DiscreteTrajectoryLowerBound/8                   -0.3814         -0.3815            60            37            60            37
BM_DiscreteTrajectoryLowerBound/64                  -0.2758         -0.2760            74            54            74            54
BM_DiscreteTrajectoryLowerBound/512                 -0.1939         -0.1942            92            74            92            74
BM_DiscreteTrajectoryLowerBound/1024                -0.1976         -0.1977            95            76            95            76
```
Note that some of the gains in `BM_DiscreteTrajectoryFind` are due to the earlier changes to `DiscreteTrajectory::end`.