@pleroy 2022-01-24
FWIW I have appended a blurb to the [design document](https://docs.google.com/document/d/1POE2fXMCMndNPzhHelkrmKAS1TvFsZ7pBOZk_9sqtJs/edit?usp=sharing) explaining why we need empty segments and how they work.